# App/iom-backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IOM Backend API",
    "description": "API documentation for Icepik Octo Manager (IOM) backend system",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/print-jobs": {
      "get": {
        "tags": ["Print Jobs"],
        "summary": "Get all print jobs for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of print jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Print Jobs"],
        "summary": "Create a new print job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": {
                    "type": "string"
                  },
                  "printer": {
                    "type": "string"
                  }
                },
                "required": ["filename", "printer"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Print job created"
          }
        }
      }
    },
...
  }
}