# App/iom-backend/routes/printJobs.js
// routes/printJobs.js
const express = require("express");
const router = express.Router();

const auth = require("../middleware/authMiddleware");
const adminOnly = require("../middleware/adminMiddleware");

const PrintJob = require("../models/PrintJob");
const { createPrintJob } = require("../controllers/printJobController");
const { logEvent } = require("../services/analyticsService");

// === Create a new print job (optional — separate from upload)
router.post("/", auth, createPrintJob);

// === User's print history
router.get("/history", auth, async (req, res) => {
  try {
    const { status, printer, page = 1, limit = 10 } = req.query;
    const filter = { userId: req.user.id };
    if (status) filter.status = status;
    if (printer) filter.printer = printer;

    const skip = (parseInt(page) - 1) * parseInt(limit);
    const history = await PrintJob.find(filter)
      .populate("modelFile")
      .sort({ createdAt: -1 })
      .skip(skip)
      .limit(parseInt(limit));

    res.json(history);
  } catch (err) {
    console.error(err);
    res.status(500).json({ message: "Failed to load print history." });
  }
});

// === Get a single print job (admin or owner)
router.get("/:id", auth, async (req, res) => {
  try {
    const job = await PrintJob.findById(req.params.id)
      .populate("modelFile")
      .populate("userId", "username email");
    if (!job) return res.status(404).json({ message: "Print job not found" });

    const isOwner = job.userId._id.toString() === req.user.id;
    const isAdmin = req.user.subscriptionTier === "admin" || req.user.isAdmin;

    if (!isOwner && !isAdmin) {
      return res.status(403).json({ message: "Unauthorized" });
    }

    res.json(job);
  } catch (err) {
    console.error(err);
    res.status(500).json({ message: "Failed to fetch job details." });
  }
});

// === Admin: all print jobs with filters
router.get("/", auth, adminOnly, async (req, res) => {
  try {
    const { status, printer, userId, page = 1, limit = 20 } = req.query;
    const filter = {};
    if (status) filter.status = status;
    if (printer) filter.printer = printer;
    if (userId) filter.userId = userId;

    const skip = (parseInt(page) - 1) * parseInt(limit);
    const jobs = await PrintJob.find(filter)
      .populate("userId", "username email")
      .populate("modelFile")
      .sort({ createdAt: -1 })
      .skip(skip)
      .limit(parseInt(limit));

    res.json(jobs);
  } catch (err) {
    console.error(err);
    res.status(500).json({ message: "Failed to load print jobs." });
  }
});

// === Admin: delete print job
router.delete("/:id", auth, adminOnly, async (req, res) => {
  try {
    const job = await PrintJob.findById(req.params.id);
    if (!job) return res.status(404).json({ message: "Print job not found" });

    await PrintJob.deleteOne({ _id: req.params.id });

    await logEvent(req.user.id, "admin_deleted_print_job", {
      jobId: job._id,
      filename: job.filename,
      userId: job.userId,
    });

    res.json({ message: "🗑️ Print job deleted successfully." });
  } catch (err) {
    console.error("❌ Failed to delete print job:", err.message);
    res.status(500).json({ message: "Internal server error." });
  }
});

module.exports = router;